enable stricter conversion checking for gcc < 4.9, fix code

+++ src/battle.c
@@ -1695,7 +1695,7 @@ void do_combatmagic(battle * b, combatmagic_t was)
 
             level = eff_skill(mage, SK_MAGIC, r);
             if (level > 0) {
-                float power;
+                double power;
                 const spell *sp;
                 const struct locale *lang = mage->faction->locale;
                 order *ord;
@@ -2497,14 +2497,14 @@ troop select_ally(fighter * af, int minrow, int maxrow, int allytype)
 static int loot_quota(const unit * src, const unit * dst,
     const item_type * type, int n)
 {
-    static float divisor = -1;
+    static double divisor = -1;
     if (dst && src && src->faction != dst->faction) {
         if (divisor < 0) {
             divisor = get_param_flt(global.parameters, "rules.items.loot_divisor", 1);
             assert(divisor == 0 || divisor >= 1);
         }
         if (divisor >= 1) {
-            double r = (float)n / divisor;
+            double r = n / divisor;
             int x = (int)r;
 
             r = r - x;
