+++ src/spells/shipcurse.h
@@ -29,8 +29,8 @@ extern "C" {
         const struct curse *c, int self);
     void register_shipcurse(void);
     struct curse *shipcurse_flyingship(struct ship *sh, struct unit *mage,
-        float power, int duration);
-    int levitate_ship(struct ship *sh, struct unit *mage, float power,
+        double power, int duration);
+    int levitate_ship(struct ship *sh, struct unit *mage, double power,
         int duration);
 
 #ifdef __cplusplus
